requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
chromadb>=0.4.0
pypdf2>=3.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
pandas>=1.5.0
plotly>=5.0.0
requests>=2.28.0
